feat(Mapping): automatically remap duplicate
This commit fix @vixyd comment,
There is no more warning for duplicate during startup,
Duplicates found at runtime are reported on the message section
(background).

The only counter intuitive case would be if user-defined and automatic
key maps are mixed (unlikely). In this case, if a user-defined key map
is already in use by a previous automatically assigned item, the
user-defined key map is ignored (and a warning is raised).
This would be costly to check and is now documented.